feat: move image input into image dialog and improve file selection handling

+++ backend/src/job_controller/state.rs
@@ -11,7 +11,7 @@ pub struct JobsState {
 #[derive(Clone, Debug, Serialize)]
 pub enum JobStatus {
     Pending,
-    InProgress(u8),
+    InProgress(u32),
     Completed(String),
     Failed(String),
 }
